KAFKA-9692: Fix flaky test
The assertion failure in KAFKA-9692 can only have been caused by the call to
ReassignPartitionsCommand.executeAssignment returning async wrt the controller
observing the reassignment in ZK (which is does) and the call to
waitUntilTrue returning because the listPartitionReassignments() observed the
initial reassignment still in progress.

Fix it by waiting until the assignments cause by the call to
ReassignPartitionsCommand.executeAssignment are observed via the AdminClient
before making the assertions.