KAFKA-9401: Change locking in FetchSessionCache to avoid contention
Avoid using intrinsic locks on the whole cache and sessions by using
concurrent collections and a `ReentrantLock` on the session itself.
Update the intrinsic locking of sessions to use this explict `ReentrantLock`.

It is necessary to account for the fact that the size of a `ConcurrentHashMap`
is at best an approximation. If we used the size to limit the number of
entries we might end up exceeding `maxEntries`. We use a `Semaphore` to
count the number of free slots in the cache, `tryAcquire` succeeds when
there are free slots. Because a successful `tryEvict()` is always accomanied
by the addition of a new session, we don't `release` the semaphore when a
slot is freed by eviction.

No lock is held between `newSessionId()` and the addition of the session
to `sessions` so it is possible for different threads to generate the same id.
To avoid this we generate a new session if the session id if the generated
one has appeared in the map since it was generated.

On the `tryEvict()` path it is necessary to double check that sessions we're
going to evict are indeed still in the cache due to `tryEvict()`-`tryEvict()`
races.

In `touch()` we add the session under the new key prior to removing its entry
with old key so that `tryEvict()` can't observe the `evictableBy*` map
without the session.

fixup